qb-server-agent/requirements.txt
Flask==3.0.0
flask-cors==6.0.1
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
requests>=2.30.0
APScheduler==3.11.0
python-json-logger==4.0.0
waitress==3.0.2
SQLAlchemy==2.0.25